Do not render children without anchors
This avoids a problem that can occur when a view is registered without
an anchor and then manually rendered by the developer later (in
`afterRender` usually).

If an un-anchored view is rendered by backbone layout, the un-anchored
view will not be appened to the DOM. However, the events associated with
that view will be delegated. If the developer later re-renders the view
and appends it to the DOM manually a bug manifests that causes events on
the un-anchored view to be overwritten and/or become unresponsive.

This commit addresses that issue by ensuring that views are only
rendered when they have an anchor point.